+++ pytorch_segmentation_models_trainer/custom_callbacks/training_callbacks.py
@@ -33,13 +33,16 @@ class WarmupCallback(pl.callbacks.base.Callback):
         super().__init__()
         self.warmup_epochs = warmup_epochs
         self.warmed_up = False
+        self.trainable = False
 
     def on_init_end(self, trainer):
         print(f"\nWarmupCallback initialization at epoch {trainer.current_epoch}.\n")
         if trainer.current_epoch > self.warmup_epochs - 1:
             self.warmed_up = True
 
-    def on_train_start(self, trainer, pl_module):
+    def on_train_epoch_start(self, trainer, pl_module):
+        if self.warmed_up:
+            return
         if not self.warmed_up:
             print(
                 f"\nModel will warm up for {self.warmup_epochs} "
@@ -47,7 +50,8 @@ class WarmupCallback(pl.callbacks.base.Callback):
             )
             pl_module.set_encoder_trainable(trainable=False)
 
-    def on_train_end(self, trainer, pl_module):
+    def on_train_epoch_end(self, trainer, pl_module, unused):
+        
         if self.warmed_up:
             return
         if trainer.current_epoch >= self.warmup_epochs - 1: