+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Version 0.5.1
+
+- Bug fixes when inputs are RGBA images;
+- Bug fixes on frame field model with models other than U-Net;
+- Bug fixes on FrameFieldResultCallback (all black image fixed).
+
 # Version 0.5.0
 
 - Added frame field training image visualization callback.